Indecisive Help
Indecisive Help is a program/tool that helps a user make a decision that they are 50/50 about. Using a simple but fun graphical user interface and object-oriented programing, this program is easy for the user to use and it is even easier for the user to make the decision that they were on the edge about.