Order Up
Order Up is a program made with HTML, CSS and JavaScript. It replicates what it would look like to order a pizza online. Including:  Pizza size, choice of toppings, and multiple ways to get a discount on the pizza the user just ordered while the final price is continuously updated depending on what the user decides to pick.